+++ onmt/train/init.lua
@@ -1,7 +1,7 @@
 local train = {}
 
 train.Trainer = require('onmt.train.Trainer')
-train.Checkpoint = require('onmt.train.Checkpoint')
+train.Saver = require('onmt.train.Saver')
 train.EpochState = require('onmt.train.EpochState')
 train.Optim = require('onmt.train.Optim')
